refactor: P1 polish — presentation names, warnings, handler param
- di/presentation: 4 modules renamed to domain-aligned helpers (surgeryMedication, imagingLamenessVaccination, icsiReproEvent, dentistryFarrier), Patient unchanged
- warnings: 0 commonMain warnings — month.number/day, expect/actual @file:Suppress blank line + build flag -Xexpect-actual-classes, readLine, Glass .toFloat removed, CloudKit drain suppress, 20 handlers r→record
- oracle gate P1 final PASS (blank line fix, pilot deferred)

+++ shared/src/commonMain/kotlin/com/github/rodrigotimoteo/animally/presentation/common/glass/GlassTopAppBar.kt
@@ -100,9 +100,9 @@ private fun isDarkTheme(): Boolean {
 }
 
 private fun Color.luminance(): Float {
-    val r = red.toFloat()
-    val g = green.toFloat()
-    val b = blue.toFloat()
+    val r = red
+    val g = green
+    val b = blue
     return LUMINANCE_RED_COEFFICIENT * r + LUMINANCE_GREEN_COEFFICIENT * g + LUMINANCE_BLUE_COEFFICIENT * b
 }
